+++ admin/class-yoast-form.php
@@ -1,7 +1,6 @@
 <?php
 /**
- * @package    WPSEO
- * @subpackage Admin
+ * @package WPSEO\Admin
  */
 
 /**
@@ -40,9 +39,9 @@ class Yoast_Form {
 	/**
 	 * Generates the header for admin pages
 	 *
-	 * @param bool   $form           Whether or not the form start tag should be included.
-	 * @param string $option         The short name of the option to use for the current page.
-	 * @param bool   $contains_files Whether the form should allow for file uploads.
+	 * @param bool   $form             Whether or not the form start tag should be included.
+	 * @param string $option           The short name of the option to use for the current page.
+	 * @param bool   $contains_files   Whether the form should allow for file uploads.
 	 * @param bool   $option_long_name
 	 */
 	public function admin_header( $form = true, $option = 'wpseo', $contains_files = false, $option_long_name = false ) {
@@ -77,8 +76,8 @@ class Yoast_Form {
 	 * @param string $option_name
 	 */
 	public function set_option( $option_name ) {
-		$this->option_name  = $option_name;
-		$this->options = $this->get_option();
+		$this->option_name = $option_name;
+		$this->options     = $this->get_option();
 	}
 
 	/**
@@ -92,6 +91,7 @@ class Yoast_Form {
 		if ( is_network_admin() ) {
 			return get_site_option( $this->option_name );
 		}
+
 		return get_option( $this->option_name );
 	}
 
@@ -289,14 +289,24 @@ class Yoast_Form {
 	/**
 	 * Create a Text input field.
 	 *
-	 * @param string $var   The variable within the option to create the text input field for.
-	 * @param string $label The label to show for the variable.
+	 * @param string       $var   The variable within the option to create the text input field for.
+	 * @param string       $label The label to show for the variable.
+	 * @param array|string $attr  Extra class to add to the input field
 	 */
-	public function textinput( $var, $label ) {
-		$val = ( isset( $this->options[ $var ] ) ) ? $this->options[ $var ] : '';
+	public function textinput( $var, $label, $attr = array() ) {
+		if ( ! is_array( $attr ) ) {
+			$attr = array(
+				'class' => $attr,
+			);
+		}
+		$attr = wp_parse_args( $attr, array(
+			'placeholder' => '',
+			'class'       => '',
+		) );
+		$val  = ( isset( $this->options[ $var ] ) ) ? $this->options[ $var ] : '';
 
 		$this->label( $label . ':', array( 'for' => $var ) );
-		echo '<input class="textinput" type="text" id="', esc_attr( $var ), '" name="', esc_attr( $this->option_name ), '[', esc_attr( $var ), ']" value="', esc_attr( $val ), '"/>', '<br class="clear" />';
+		echo '<input class="textinput ' . esc_attr( $attr['class'] ) . ' " placeholder="' . esc_attr( $attr['placeholder'] ) . '" type="text" id="', esc_attr( $var ), '" name="', esc_attr( $this->option_name ), '[', esc_attr( $var ), ']" value="', esc_attr( $val ), '"/>', '<br class="clear" />';
 	}
 
 	/**