+++ premium/classes/class-page-redirect.php
@@ -1,15 +1,14 @@
 <?php
+
 /**
  * @package Premium\Redirect
  */
-
 class WPSEO_Page_Redirect {
 
 	/**
 	 * Function that outputs the redirect page
 	 */
 	public static function display() {
-		global $wpseo_admin_pages;
 
 		// Check if there's an old URL set
 		$old_url = '';
@@ -21,17 +20,21 @@ class WPSEO_Page_Redirect {
 		$redirect_types = WPSEO_Redirect_Manager::get_redirect_types();
 
 		// Admin header
-		Yoast_Form::get_instance()->admin_header( false, 'yoast_wpseo_redirects_options', 'wpseo_redirects' );
+		Yoast_Form::get_instance()->admin_header( false, 'wpseo_redirects', false, 'yoast_wpseo_redirects_options');
 		?>
 		<h2 class="nav-tab-wrapper" id="wpseo-tabs">
-			<a class="nav-tab" id="tab-url-tab" href="#top#tab-url"><?php _e( 'Redirects', 'wordpress-seo-premium' ); ?></a>
-			<a class="nav-tab" id="tab-regex-tab" href="#top#tab-regex"><?php _e( 'Regex Redirects', 'wordpress-seo-premium' ); ?></a>
-			<a class="nav-tab" id="settings-tab" href="#top#settings"><?php _e( 'Settings', 'wordpress-seo-premium' ); ?></a>
+			<a class="nav-tab" id="tab-url-tab"
+			   href="#top#tab-url"><?php _e( 'Redirects', 'wordpress-seo-premium' ); ?></a>
+			<a class="nav-tab" id="tab-regex-tab"
+			   href="#top#tab-regex"><?php _e( 'Regex Redirects', 'wordpress-seo-premium' ); ?></a>
+			<a class="nav-tab" id="settings-tab"
+			   href="#top#settings"><?php _e( 'Settings', 'wordpress-seo-premium' ); ?></a>
 		</h2>
 
 		<div class="tabwrapper>">
 			<div id="tab-url" class="wpseotab redirect-table-tab">
 				<?php
+
 				// Add new redirect HTML
 				echo "<form class='wpseo-new-redirect-form' method='post'>\n";
 				echo "<div class='wpseo_redirects_new'>\n";
@@ -60,7 +63,10 @@ class WPSEO_Page_Redirect {
 
 				echo "</select>" . PHP_EOL;
 
-				echo '<img src="' . plugins_url( 'images/question-mark.png', WPSEO_FILE ) . '" class="yoast_help" style="margin:17px 0 0 5px" alt="' . __( sprintf( 'The redirect type is de HTTP response code send to the browser telling the browser what type of redirect is served.<br/><br/>Read <a href=\'%s\'>this page</a> for more info.', 'http://kb.yoast.com/article/121-redirect-types/#utm_source=wordpress-seo-premium-redirects&amp;utm_medium=inline-help&amp;utm_campaign=redirect-types' ), 'wordpress-seo-premium' ) . '">';
+				echo '<br />';
+				echo '<br />';
+
+				echo '<p class="label desc description">' . sprintf( __( 'The redirect type is the HTTP response code send to the browser telling the browser what type of redirect is served.<br/><br/>Read <a href=\'%s\'>this page</a> for more info.', 'wordpress-seo-premium' ), 'http://kb.yoast.com/article/121-redirect-types/#utm_source=wordpress-seo-premium-redirects&amp;utm_medium=inline-help&amp;utm_campaign=redirect-types' ) . '</p>';
 
 				echo "<br class='clear'/>\n";
 
@@ -120,7 +126,10 @@ class WPSEO_Page_Redirect {
 
 				echo "</select>" . PHP_EOL;
 
-				echo '<img src="' . plugins_url( 'images/question-mark.png', WPSEO_FILE ) . '" class="yoast_help" style="margin:17px 0 0 5px" alt="' . __( sprintf( 'The redirect type is de HTTP response code send to the browser telling the browser what type of redirect is served.<br/><br/>Read <a href=\'%s\'>this page</a> for more info.', 'http://kb.yoast.com/article/121-redirect-types/#utm_source=wordpress-seo-premium-redirects&amp;utm_medium=inline-help&amp;utm_campaign=redirect-types' ), 'wordpress-seo-premium' ) . '">';
+				echo '<br />';
+				echo '<br />';
+
+				echo '<p class="label desc description">' . sprintf( __( 'The redirect type is the HTTP response code send to the browser telling the browser what type of redirect is served.<br/><br/>Read <a href=\'%s\'>this page</a> for more info.', 'wordpress-seo-premium' ), 'http://kb.yoast.com/article/121-redirect-types/#utm_source=wordpress-seo-premium-redirects&amp;utm_medium=inline-help&amp;utm_campaign=redirect-types' ) . ' </p>';
 
 				echo "<br class='clear'/>\n";
 
@@ -166,23 +175,27 @@ class WPSEO_Page_Redirect {
 								echo '<p>' . __( "As you're on Apache, you should add the following include to the website httpd config file:", 'wordpress-seo-premium' ) . '</p>';
 								echo '<pre>Include ' . WPSEO_Redirect_File_Manager::get_file_path() . '</pre>';
 								echo '</div>';
-							} else {
+							}
+							else {
 								$file_write_error = true;
 							}
-						} else {
+						}
+						else {
 							if ( ! is_writable( WPSEO_Redirect_File_Manager::get_htaccess_file_path() ) ) {
 								/* translators: %s: '.htaccess' file name */
 								echo "<div class='error'><p><b>" . sprintf( __( 'We\'re unable to save the redirects to your %s file. Please make the file writable.', 'wordpress-seo-premium' ), '<code>.htaccess</code>' ) . "</b></p></div>\n";
 							}
 						}
 
-					} else if ( WPSEO_Utils::is_nginx() ) {
+					}
+					else if ( WPSEO_Utils::is_nginx() ) {
 						if ( file_exists( WPSEO_Redirect_File_Manager::get_file_path() ) ) {
 							echo '<div style="margin: 5px 0; padding: 3px 10px; background-color: #ffffe0; border: 1px solid #E6DB55; border-radius: 3px">';
 							echo '<p>' . __( 'As you\'re on Nginx, you should add the following include to the NGINX config file:', 'wordpress-seo-premium' ) . '</p>';
 							echo '<pre>include ' . WPSEO_Redirect_File_Manager::get_file_path() . ';</pre>';
 							echo '</div>';
-						} else {
+						}
+						else {
 							$file_write_error = true;
 						}
 					}
@@ -201,19 +214,21 @@ class WPSEO_Page_Redirect {
 				<form action="<?php echo admin_url( 'options.php' ); ?>" method="post">
 					<?php
 					settings_fields( 'yoast_wpseo_redirect_options' );
-					$wpseo_admin_pages->currentoption = 'wpseo_redirect';
+
+					Yoast_Form::get_instance()->set_option('wpseo_redirect');
 
 
 					if ( WPSEO_Utils::is_apache() ) {
 
 						echo Yoast_Form::get_instance()->checkbox( 'disable_php_redirect', __( 'Disable PHP redirects', 'wordpress-seo-premium' ) );
+
 						/* translators: 1: '.htaccess' file name */
 						echo '<p class="desc">' . sprintf( __( 'Write redirects to the %1$s file. Make sure the %1$s file is writable.', 'wordpress-seo-premium' ), '<code>.htacces</code>' ) . '</p>';
 
 						echo Yoast_Form::get_instance()->checkbox( 'separate_file', __( 'Generate a separate redirect file', 'wordpress-seo-premium' ) );
+
 						/* translators: %s: '.htaccess' file name */
 						echo '<p class="desc">' . sprintf( __( 'By default we write the redirects to your %s file, check this if you want the redirects written to a separate file. Only check this option if you know what you are doing!', 'wordpress-seo-premium' ), '<code>.htaccess</code>' ) . '</p>';
-
 					} else {
 						echo Yoast_Form::get_instance()->checkbox( 'disable_php_redirect', __( 'Disable PHP redirects', 'wordpress-seo-premium' ) );
 						echo '<p class="desc">' . __( "WordPress SEO will generate redirect files that can be included in your website configuration. You can disable PHP redirect if this is done correctly. Only check this option if you know what you are doing!", 'wordpress-seo-premium' ) . '</p>';
@@ -221,7 +236,8 @@ class WPSEO_Page_Redirect {
 
 					?>
 					<p class="submit">
-						<input type="submit" name="submit" id="submit" class="button button-primary" value="<?php _e( 'Save Changes', 'wordpress-seo-premium' ); ?>">
+						<input type="submit" name="submit" id="submit" class="button button-primary"
+						       value="<?php _e( 'Save Changes', 'wordpress-seo-premium' ); ?>">
 					</p>
 				</form>
 			</div>
@@ -245,8 +261,8 @@ class WPSEO_Page_Redirect {
 	 */
 	public static function page_scripts() {
 		wp_enqueue_script( 'jquery-qtip', plugins_url( 'js/jquery.qtip.min.js', WPSEO_FILE ), array( 'jquery' ), '1.0.0-RC3', true );
-		wp_enqueue_script( 'wpseo-premium-yoast-overlay', plugin_dir_url( WPSEO_PREMIUM_FILE ) . '/assets/js/wpseo-premium-yoast-overlay.js', array( 'jquery' ), '1.0.0' );
-		wp_enqueue_script( 'wp-seo-premium-admin-redirects', plugin_dir_url( WPSEO_PREMIUM_FILE ) . '/assets/js/wp-seo-premium-admin-redirects.js', array( 'jquery' ), '1.0.0' );
+		wp_enqueue_script( 'wpseo-premium-yoast-overlay', plugin_dir_url( WPSEO_PREMIUM_FILE ) . 'assets/js/wpseo-premium-yoast-overlay' . WPSEO_CSSJS_SUFFIX . '.js', array( 'jquery' ), '1.0.0' );
+		wp_enqueue_script( 'wp-seo-premium-admin-redirects', plugin_dir_url( WPSEO_PREMIUM_FILE ) . 'assets/js/wp-seo-premium-admin-redirects' . WPSEO_CSSJS_SUFFIX . '.js', array( 'jquery' ), '1.0.0' );
 		wp_localize_script( 'wp-seo-premium-admin-redirects', 'wpseo_premium_strings', WPSEO_Premium_Javascript_Strings::strings() );
 
 		add_screen_option( 'per_page', array(