+++ premium/classes/class-post-watcher.php
@@ -12,6 +12,20 @@ class WPSEO_Post_Watcher extends WPSEO_Watcher {
 	protected $watch_type = 'post';
 
 	/**
+	 * Constructor of class
+	 */
+	public function __construct() {
+		add_action( 'admin_enqueue_scripts', array( $this, 'page_scripts' ) );
+	}
+
+	/**
+	 * Load needed js file
+	 */
+	public function page_scripts() {
+		wp_enqueue_script( 'wp-seo-premium-quickedit-notification', plugin_dir_url( WPSEO_PREMIUM_FILE ) . 'assets/js/wp-seo-premium-quickedit-notification.js', array( 'jquery' ) );
+	}
+
+	/**
 	 * Add an extra field to post edit screen so we know the old url in the 'post_updated' hook
 	 *
 	 * @param $post
@@ -34,27 +48,19 @@ class WPSEO_Post_Watcher extends WPSEO_Watcher {
 	 * @param $post_before
 	 */
 	public function detect_slug_change( $post_id, $post, $post_before ) {
+		$old_url = $this->get_old_url( $post, $post_before );
 
-		if ( !isset( $_POST['wpseo_old_url'] ) ) {
+		if ( !$old_url ) {
 			return;
 		}
 
 		// Get the new URL
 		$new_url = $this->get_target_url( $post_id );
 
-		// Get the old URL
-		$old_url = esc_url( $_POST['wpseo_old_url'] );
-
 		// Check if we should create a redirect
 		if ( in_array( $post->post_status, array( 'publish', 'static' ) ) && $this->should_create_redirect( $old_url, $new_url ) ) {
-			// Format the message
-			$message = sprintf( __( 'WordPress SEO Premium created a <a href="%s">redirect</a> from the old post URL to the new post URL. <a href="%s">Click here to undo this</a>.', 'wordpress-seo-premium' ), $this->admin_redirect_url( $old_url ), $this->javascript_undo_redirect( $old_url ) );
-
-			$this->create_redirect( $old_url, $new_url );
-
-			$this->create_notification( $message, 'slug_change' );
+			$this->set_notification( $old_url, $new_url );
 		}
-
 	}
 
 	/**
@@ -167,6 +173,41 @@ class WPSEO_Post_Watcher extends WPSEO_Watcher {
 		$url = $url['path'];
 
 		return $url;
+	}
+
+	/**
+	 * Get the old url
+	 *
+	 * @param $post
+	 * @param $post_before
+	 *
+	 * @return bool|string
+	 */
+	protected function get_old_url( $post, $post_before ) {
+		if ( !isset( $_POST['wpseo_old_url'] ) ) {
+			// Check if request is inline action and new slug is not old slug, if so set wpseo_old_url
+			if ( ! empty( $_POST['action'] ) && $_POST['action'] === 'inline-save' && $post->post_name !== $post_before->post_name ) {
+				return '/' . $post_before->post_name . '/';
+			}
+			return false;
+		}
+
+		return esc_url( $_POST['wpseo_old_url'] );
+	}
+
+	/**
+	 * Display notification
+	 *
+	 * @param $old_url
+	 * @param $new_url
+	 */
+	protected function set_notification( $old_url, $new_url ) {
+		$message = sprintf( __( "WordPress SEO Premium created a <a href='%s'>redirect</a> from the old post URL to the new post URL. <a href='%s'>Click here to undo this</a>.", 'wordpress-seo-premium' ), $this->admin_redirect_url( $old_url ), $this->javascript_undo_redirect( $old_url ) );
+
+		$this->create_redirect( $old_url, $new_url );
+
+		//Only set notification when the slug change was not saved through quick edit
+		$this->create_notification( $message, 'slug_change' );
 	}
 
 }